+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 Changes:
+- Added chat message limit setting